# api/openapi.yaml
openapi: 3.1.0
info:
  title: Farmish API
  version: 0.1.0
  summary: Farmish Ghana agricultural marketplace API.
  description: |
    Contract-first HTTP API for Farmish. This file is the source of truth:
    Go server types are generated from it (`make generate`), and requests are
    validated against it at runtime.

    Conventions:
    - JSON properties and query parameters are camelCase.
    - Money is always integer pesewas (`Money.amount`) plus `currency: GHS`.
    - Errors use one envelope: `{"error": {"code", "message", "details?"}}`.
      `code` is a stable snake_case identifier; `message` is human-readable
      and may change.
    - List endpoints take `page` (from 1) and `limit` (1-50) and return `PageMeta`.
    - Authenticated endpoints take a Firebase ID token as `Authorization: Bearer <token>`.
      Every operation requires it unless it declares `security: []`. Missing,
      expired or invalid tokens all get the same 401 `unauthorized` envelope.
    - `x-farmish-role: admin` on an operation additionally requires the admin
      role (403 `forbidden` otherwise). The server enforces exactly what the
      spec declares.
    - `x-farmish-step-up: true` on an operation requires a fresh sign-in
      (Firebase `auth_time` within 5 minutes) verified with a revocation
      check. A stale sign-in gets 401 `reauth_required` with
      `WWW-Authenticate: Bearer error="insufficient_user_authentication"`,
...
      tokens still get the generic 401 `unauthorized`.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/listings:
    get:
      tags: [listings]
      operationId: searchListings
      summary: Search and browse listings
      description: |
        Active, unexpired listings, promoted ones first (DOMAIN §6) and then
        by the requested sort. Anonymous callers are allowed; the response
        never contains a seller's contact details. Cacheable for 30 seconds,
        and a matching `If-None-Match` returns 304 with no body.
      security: []
      parameters:
        - name: q
          in: query
          description: Free text, matched against the title and description.
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: category
          in: query
          description: Category slug. A parent includes all its children.
          schema:
            type: string
            maxLength: 80
        - name: region
          in: query
          schema:
            $ref: "#/components/schemas/GhanaRegion"
        - name: district
          in: query
          schema:
            type: string
            maxLength: 80
        - name: minPrice
          in: query
          description: Lowest price, in pesewas.
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: maxPrice
          in: query
          description: Highest price, in pesewas.
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: itemState
          in: query
          schema:
            type: string
            maxLength: 40
        - name: sort
          in: query
          description: |
            `relevance` needs `q`; without it the sort falls back to `newest`.
          schema:
            type: string
            enum: [relevance, newest, price_asc, price_desc]
            default: newest
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: If-None-Match
          in: header
          description: |
            A validator from an earlier response. When it still matches, the
            server answers 304 with no body.
          schema:
            type: string
      responses:
        "200":
          description: A page of listings.
          headers:
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingSearchResult"
        "304":
          $ref: "#/components/responses/NotModified"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [listings]
      operationId: createListing
      summary: Create a listing
      description: |
        Creates a listing for the signed-in seller, who must have a seller
        profile (403 `seller_profile_required` otherwise). With
        `publish: true` the listing goes live in the same request and must be
        complete (at least one image, every required attribute).
      x-farmish-rate-limit: sensitive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: The created listing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerListing"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/listings/{slug}:
    get:
      tags: [listings]
      operationId: getPublicListing
      summary: Get one listing for a buyer
      description: |
        The public listing page: the summary plus the detail fields and the
        safe seller projection. A listing that is not active, or is expired,
        is a 404 so a stale link never shows a dead page. No phone, email or
        other contact detail appears here: the contact endpoint is the only
        way to see a seller's number, and only when the seller opted in.

        A view is counted asynchronously (`listings.count_view`), so this
        stays a cheap read. Cacheable for 60 seconds.
      security: []
      parameters:
        - $ref: "#/components/parameters/ListingSlug"
        - name: If-None-Match
          in: header
          description: |
            A validator from an earlier response. When it still matches, the
            server answers 304 with no body.
          schema:
            type: string
      responses:
        "200":
          description: The listing.
          headers:
            Cache-Control:
              $ref: "#/components/headers/Cache-Control"
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "304":
          $ref: "#/components/responses/NotModified"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/listings/{id}/contact:
    post:
      tags: [listings]
      operationId: revealListingContact
      summary: Show a seller's contact details
      description: |
        Returns only the fields the seller opted into: `phone` when
        `showPhone` is set, `whatsapp` when `showWhatsapp` is. A seller who
        opted out of both gets `{}`. Asking about your own listing is a 403,
        and a listing that is not browsable is a 404. Every reveal is counted
        in the listing's `contactCount`.
      x-farmish-rate-limit: sensitive
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: The opted-in contact details; absent fields are omitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingContact"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token.
  parameters:
    ListingId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ListingSlug:
      name: slug
      in: path
      required: true
      description: The listing's public slug.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Page size.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 50
        default: 20
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotModified:
      description: |
        The client's `If-None-Match` still matches, so the cached copy stands.
        Sent with no body.
      headers:
        Cache-Control:
          $ref: "#/components/headers/Cache-Control"
        ETag:
          $ref: "#/components/headers/ETag"
    TooManyRequests:
      description: Rate limit exceeded. Retry after `Retry-After` seconds.
      headers:
        Retry-After:
          description: Seconds until a request will be allowed.
          schema:
            type: integer
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  headers:
    Cache-Control:
      description: How long a shared cache may reuse the response.
      schema:
        type: string
    ETag:
      description: A weak validator over the exact response body.
      schema:
        type: string
    X-RateLimit-Limit:
      description: Burst capacity of the limit that applied.
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: Requests left before limiting.
      schema:
        type: integer
    X-RateLimit-Reset:
      description: Seconds until the limit is fully replenished.
      schema:
        type: integer
  schemas:
...
    Error:
      type: object
      description: The single error envelope used by every endpoint.
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    GhanaRegion:
      type: string
      description: One of Ghana's 16 regions.
      enum: [Ahafo, Ashanti, Bono, Bono East, Central, Eastern, Greater Accra, North East, Northern, Oti, Savannah, Upper East, Upper West, Volta, Western, Western North]
    ListingCategoryRef:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
        name:
          type: string
    ListingSellerRef:
      type: object
      description: Just enough seller to render a result row.
      required: [name, verified]
      properties:
        name:
          type: string
          description: The seller's business name.
        verified:
          type: boolean
    ListingPromotion:
      type: object
      description: A running promotion. Only the tier is public; the rank is internal.
      required: [tier]
      properties:
        tier:
          type: string
          enum: [top, vip, diamond, enterprise]
    ListingSummary:
      type: object
      description: One row of public search or browse.
      required: [id, slug, title, price, unit, region, district, itemState, category, promoted, seller, publishedAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        price:
          $ref: "#/components/schemas/Money"
        unit:
          type: string
          description: The selling unit, e.g. `kg`, `bags`, `heads`.
        region:
          $ref: "#/components/schemas/GhanaRegion"
        district:
          type: string
        itemState:
          type: string
          description: The item's condition or grade, e.g. `grade_a`, `used`.
        category:
          $ref: "#/components/schemas/ListingCategoryRef"
        coverImageUrl:
          type: string
          description: The first image, if the seller added one.
        promoted:
          oneOf:
            - $ref: "#/components/schemas/ListingPromotion"
            - type: "null"
          description: Null when the listing is not promoted.
        seller:
          $ref: "#/components/schemas/ListingSellerRef"
        publishedAt:
          type: string
          format: date-time
    ListingImage:
      type: object
      required: [url, order]
      properties:
        url:
          type: string
        order:
          type: integer
          format: int32
          minimum: 0
    ListingAttribute:
      type: object
      required: [key, label, value]
      properties:
        key:
          type: string
        label:
          type: string
          description: The category's display label for this attribute.
        value:
          type: string
    ListingDeliveryOptions:
      type: object
      required: [pickup, sellerDelivery]
      properties:
        pickup:
          type: boolean
        sellerDelivery:
          type: boolean
        fee:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: The seller's delivery fee, when they offer delivery.
    ListingDetail:
      type: object
      description: |
        The public listing page. It repeats the `ListingSummary` fields
        because `seller` is the full safe projection here, then adds the
        detail a buyer needs. It never carries a phone number, an email
        address or any other contact detail: those come from the contact
        endpoint, and only when the seller opted in.
      required: [id, slug, title, price, unit, region, district, itemState, category, promoted, seller, publishedAt, description, quantityAvailable, minOrderQty, isNegotiable, deliveryOptions, images, attributes, expiresAt, favoriteCount]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        price:
          $ref: "#/components/schemas/Money"
        unit:
          type: string
        region:
          $ref: "#/components/schemas/GhanaRegion"
        district:
          type: string
        itemState:
          type: string
        category:
          $ref: "#/components/schemas/ListingCategoryRef"
        coverImageUrl:
          type: string
        promoted:
          oneOf:
            - $ref: "#/components/schemas/ListingPromotion"
            - type: "null"
        seller:
          $ref: "#/components/schemas/PublicSeller"
        publishedAt:
          type: string
          format: date-time
        description:
          type: string
        quantityAvailable:
          type: integer
          format: int32
          minimum: 0
        minOrderQty:
          type: integer
          format: int32
          minimum: 1
        isNegotiable:
          type: boolean
        area:
          type: string
          description: The seller's area or landmark, when given.
        deliveryOptions:
          $ref: "#/components/schemas/ListingDeliveryOptions"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ListingImage"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ListingAttribute"
        expiresAt:
          type: string
          format: date-time
        favoriteCount:
          type: integer
          format: int32
          minimum: 0
    ListingSearchResult:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ListingSummary"
        meta:
          $ref: "#/components/schemas/PageMeta"
    ListingContact:
      type: object
      description: |
        Only the fields the seller opted into. Both are omitted when the
        seller shares neither, which is an empty object rather than an error.
      properties:
        phone:
          type: string
          description: E.164, present only when `showPhone` is set.
        whatsapp:
          type: string
          description: E.164, present only when `showWhatsapp` is set.
    PublicSeller:
      type: object
      description: Safe seller projection for anonymous callers.
      required: [userId, businessName, region, district, verified, memberSince, rating]
      properties:
        userId:
          type: string
          format: uuid
        businessName:
          type: string
        region:
          type: string
        district:
          type: string
        bio:
          type: string
        verified:
          type: boolean
          description: True when the seller is verified.
        memberSince:
          type: string
          format: date-time
          description: When the seller's user account was created.
        rating:
          $ref: "#/components/schemas/SellerRating"
...
    ListingDelivery:
      type: object
      description: Delivery options. At least one must be true; a seller-delivery fee is required exactly when sellerDelivery is on.
      additionalProperties: false
      required: [pickup, sellerDelivery]
      properties:
        pickup:
          type: boolean
        sellerDelivery:
          type: boolean
        sellerDeliveryFee:
          $ref: "#/components/schemas/Money"
    CreateListingRequest:
      type: object
      additionalProperties: false
      required: [categorySlug, title, description, price, unit, quantityAvailable, itemState, region, district, deliveryOptions]
      properties:
        categorySlug:
          type: string
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
          minLength: 20
          maxLength: 5000
        price:
          $ref: "#/components/schemas/Money"
          description: Price per unit, in pesewas. Currency must be GHS.
        unit:
          type: string
          description: Must be one of the category's units.
        quantityAvailable:
          type: integer
          format: int32
          minimum: 0
        minOrderQty:
          type: integer
          format: int32
          minimum: 1
          default: 1
        isNegotiable:
          type: boolean
          default: true
        itemState:
          type: string
          description: Must be one of the category's item states.
        region:
          type: string
          enum: [Ahafo, Ashanti, Bono, Bono East, Central, Eastern, Greater Accra, North East, Northern, Oti, Savannah, Upper East, Upper West, Volta, Western, Western North]
        district:
          type: string
          minLength: 2
          maxLength: 80
        area:
          type: string
          maxLength: 120
        deliveryOptions:
          $ref: "#/components/schemas/ListingDelivery"
        attributes:
          type: object
          additionalProperties:
            type: string
          description: Attribute values keyed by the category's attribute keys.
        imageMediaIds:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            format: uuid
          description: Media ids from POST /v1/media/upload-url, in display order.
        publish:
          type: boolean
          default: false
          description: Publish immediately (the listing must then be complete).
...
    SellerListingImage:
      type: object
      required: [mediaId, url, sortOrder]
      properties:
        mediaId:
          type: string
          format: uuid
        url:
          type: string
        sortOrder:
          type: integer
          format: int32
    SellerListingAttribute:
      type: object
      required: [key, value]
      properties:
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [text, number, boolean, select, date]
        required:
          type: boolean
    SellerListing:
      type: object
      description: The owner's view of a listing. Never exposes the seller's contact details.
      required: [id, categorySlug, title, slug, description, price, unit, quantityAvailable, minOrderQty, isNegotiable, itemState, status, region, district, offersPickup, offersSellerDelivery, images, attributes, viewCount, favoriteCount, contactCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        categorySlug:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          $ref: "#/components/schemas/Money"
        unit:
          type: string
        quantityAvailable:
          type: integer
          format: int32
        minOrderQty:
          type: integer
          format: int32
        isNegotiable:
          type: boolean
        itemState:
          type: string
        status:
          type: string
          enum: [draft, active, sold, expired, archived, suspended]
        region:
          type: string
        district:
          type: string
        area:
          type: string
        offersPickup:
          type: boolean
        offersSellerDelivery:
          type: boolean
        sellerDeliveryFee:
          $ref: "#/components/schemas/Money"
        publishedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        images:
          type: array
          items:
            $ref: "#/components/schemas/SellerListingImage"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/SellerListingAttribute"
        viewCount:
          type: integer
          format: int32
        favoriteCount:
          type: integer
          format: int32
        contactCount:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Money:
      type: object
      description: An amount of money in the smallest currency unit (pesewas).
      required: [amount, currency]
      properties:
        amount:
          type: integer
          format: int64
          description: Integer pesewas. GHS 12.50 is `1250`.
          examples: [1250]
        currency:
          type: string
          enum: [GHS]
    PageMeta:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
          format: int32
          minimum: 1
        limit:
          type: integer
          format: int32
          minimum: 1
          maximum: 50
        total:
          type: integer
          format: int64
          minimum: 0
          description: Total matching items across all pages.
...
    SellerRating:
      type: object
      description: A seller aggregate over visible reviews.
      required: [average, count]
      properties:
        average:
          type: number
          description: Mean rating to one decimal; 0 with no reviews.
          examples: [4.5]
        count:
          type: integer
          format: int64
          minimum: 0
...